cambios de integer a double en campos de la db

+++ database/migrations/2021_06_07_113302_create_vp_orders_table.php
@@ -22,9 +22,9 @@ class CreateVpOrdersTable extends Migration
             $table->string('shopChannelName')->nullable();
             $table->enum('status', ['WAITING_ACCEPTANCE','PENDING', 'PROCESSING', 'SHIPPED','CANCELLED'])->default('WAITING_ACCEPTANCE');
             $table->timestamp('shippedOrderDate')->nullable();
-            $table->integer('totalPrice')->nullable();
-            $table->integer('shippingCosts')->nullable();
-            $table->integer('shippingTaxRate')->nullable();
+            $table->double('totalPrice')->nullable();
+            $table->double('shippingCosts')->nullable();
+            $table->double('shippingTaxRate')->nullable();
             $table->string('currency')->nullable();
             $table->string('requestedShippingMethod')->nullable();
             $table->string('deliveryNote')->nullable();
